requirements.txt
mlflow==2.9.2
numpy==1.23.4
pandas==1.5.1
scikit_learn==1.1.3
python-dotenv==1.0.1
torch==2.0.1
